Refactor To Make read_backup_file a Common Function
Moves the `read_backup_file` function out of the
`restore_backup_command` command and into a common location
