+++ Makefile
@@ -4,7 +4,8 @@ AR ?=ar
 OBJ_DIR =./obj
 LIB_DIR =./lib
 
-LIBS =${LIB_DIR}/libserial.a ${LIB_DIR}/libcommands.a ${LIB_DIR}/libsupport.a ${LIB_DIR}/libapi.a ${LIB_DIR}/libencryption.a
+LIBS =${LIB_DIR}/libserial.a ${LIB_DIR}/libcommands.a ${LIB_DIR}/libcommandscommon.a \
+	  ${LIB_DIR}/libsupport.a ${LIB_DIR}/libapi.a ${LIB_DIR}/libencryption.a
 EXE =bluefish
 
 CXX_FLAGS +=--std=c++17 -Wall -Wextra -pedantic -Werror -I./cdif -I./ -O3
