+++ api/api.h
@@ -9,6 +9,8 @@
 #include "either/either.h"
 
 struct Success {};
+struct MasterBlock;
+struct File;
 
 enum class APIFailureReason : uint8_t
 {
