+++ api/file.h
@@ -4,6 +4,8 @@
 
 #include "api/serialization.h"
 
+namespace bf {}
+
 struct File
 {
     std::string name;
